+++ libs/opengl/src/TRenderMatrices.cpp
@@ -9,6 +9,7 @@
 
 #include "opengl-precomp.h"	 // Precompiled header
 //
+#include <mrpt/containers/yaml.h>
 #include <mrpt/math/geometry.h>	 // crossProduct3D()
 #include <mrpt/math/ops_containers.h>  // dotProduct()
 #include <mrpt/opengl/TRenderMatrices.h>
@@ -21,6 +22,8 @@ void TRenderMatrices::computeOrthoProjectionMatrix(
 	float left, float right, float bottom, float top, float znear, float zfar)
 {
 	ASSERT_GT_(zfar, znear);
+	m_last_z_near = znear;
+	m_last_z_far = zfar;
 
 	p_matrix.setIdentity();
 
@@ -40,6 +43,9 @@ void TRenderMatrices::computeProjectionMatrix(float znear, float zfar)
 	ASSERT_GT_(zfar, .0f);
 	ASSERT_GE_(znear, .0f);
 
+	m_last_z_near = znear;
+	m_last_z_far = zfar;
+
 	if (is_projective)
 	{
 		// Was: gluPerspective()
@@ -50,7 +56,6 @@ void TRenderMatrices::computeProjectionMatrix(float znear, float zfar)
 			std::abs(aspect - std::numeric_limits<float>::epsilon()), .0f);
 
 		const float f = 1.0f / std::tan(mrpt::DEG2RAD(FOV) / 2.0f);
-
 		p_matrix.setZero();
 
 		p_matrix(0, 0) = f / aspect;
@@ -154,4 +159,32 @@ void TRenderMatrices::projectPointPixels(
 	projectPoint(x, y, z, proj_u_px, proj_v_px, proj_depth);
 	proj_u_px = (proj_u_px + 1.0f) * (viewport_width * 0.5f);
 	proj_v_px = (proj_v_px + 1.0f) * (viewport_height * 0.5f);
+}
+
+void TRenderMatrices::saveToYaml(mrpt::containers::yaml& c) const
+{
+	c = mrpt::containers::yaml::Map();
+
+	MCP_SAVE(c, initialized);
+	MCP_SAVE(c, viewport_width);
+	MCP_SAVE(c, viewport_height);
+	MCP_SAVE(c, FOV);
+	MCP_SAVE_DEG(c, azimuth);
+	MCP_SAVE_DEG(c, elev);
+	MCP_SAVE(c, eyeDistance);
+	MCP_SAVE(c, is_projective);
+
+	c["eye"] = eye.asString();
+	c["pointing"] = pointing.asString();
+	c["up"] = up.asString();
+
+	c["p_matrix"] = mrpt::containers::yaml::FromMatrix(p_matrix);
+	c["mv_matrix"] = mrpt::containers::yaml::FromMatrix(mv_matrix);
+}
+
+void TRenderMatrices::print(std::ostream& o) const
+{
+	mrpt::containers::yaml c;
+	saveToYaml(c);
+	o << c;
 }